+++ chevah/compat/tests/normal/testing/test_testcase.py
@@ -304,9 +304,6 @@ class TestTwistedTestCase(ChevahTestCase):
         deferred = threads.deferToThread(last_call)
 
         self.executeReactor()
-        # Allow the thread to settle and return.
-        time.sleep(0.01)
-        time.sleep(0.05)
         self.assertTrue(self.called)
         self.assertTrue(deferred.called)